backend/app/evaluator_agent.py
"""
Evaluator Agent - Evaluates rendered images against the original prompt.
Currently uses a stub implementation. In production, this would integrate
with a vision AI model (e.g., Claude with vision capabilities).
"""

import logging
from typing import Tuple, Dict, Any
from PIL import Image
import numpy as np

# Set up logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


def check_symmetry(image_path: str, symmetry_type: str) -> float:
    """
    Check if an image has the expected symmetry.

    Args:
        image_path: Path to the image file
        symmetry_type: Expected symmetry type ('vertical', 'horizontal', 'radial', 'none')

    Returns:
        Score from 0-10 indicating how well the symmetry matches
    """
    try:
        # Load the image
        img = Image.open(image_path).convert('L')  # Convert to grayscale
        img_array = np.array(img)

        if symmetry_type == 'none':
            return 8.0  # No symmetry requirement

        height, width = img_array.shape

        if symmetry_type == 'vertical':
            # Check left-right symmetry
            left_half = img_array[:, :width//2]
            right_half = np.fliplr(img_array[:, width//2:])

            # Make them the same size
            min_width = min(left_half.shape[1], right_half.shape[1])
            left_half = left_half[:, :min_width]
            right_half = right_half[:, :min_width]

            # Calculate similarity
            difference = np.abs(left_half.astype(float) - right_half.astype(float))
            similarity = 1.0 - (np.mean(difference) / 255.0)
            score = similarity * 10

            logger.info(f"Vertical symmetry score: {score:.2f}")
            return score

        elif symmetry_type == 'horizontal':
            # Check top-bottom symmetry
            top_half = img_array[:height//2, :]
            bottom_half = np.flipud(img_array[height//2:, :])

            # Make them the same size
            min_height = min(top_half.shape[0], bottom_half.shape[0])
            top_half = top_half[:min_height, :]
            bottom_half = bottom_half[:min_height, :]

            # Calculate similarity
            difference = np.abs(top_half.astype(float) - bottom_half.astype(float))
            similarity = 1.0 - (np.mean(difference) / 255.0)
            score = similarity * 10

            logger.info(f"Horizontal symmetry score: {score:.2f}")
            return score

        else:
            # For radial or other types, return a moderate score
            return 7.0

    except Exception as e:
        logger.error(f"Error checking symmetry: {e}")
        return 7.0  # Return moderate score on error


def evaluate_image(
    image_path: str,
    prompt_text: str,
    description: Dict[str, Any] = None,
    iteration_number: int = 1
) -> Tuple[float, str]:
    """
    Evaluate how well the rendered image matches the original prompt.

    This is a stub implementation that uses heuristics. In production,
    this would call a vision AI model (e.g., Claude with vision, GPT-4V)
    to evaluate the image.

    Args:
        image_path: Path to the rendered image
        prompt_text: Original user prompt
        description: Optional structured description from interpret_prompt
        iteration_number: Current iteration number (for progressive improvement)

    Returns:
        Tuple of (overall_score, feedback_text)
        - overall_score: Float from 0-10 indicating quality
        - feedback_text: String with specific feedback for improvement
    """
    logger.info(f"Evaluating image (iteration {iteration_number}): {image_path}")

    # NOTE: This is a stub implementation for demonstration purposes.
    # In production, this would integrate with a vision AI model.

    # Placeholder evaluation logic
    # We simulate progressive improvement across iterations

    try:
        # Check if image exists and is valid
        img = Image.open(image_path)
        width, height = img.size

        logger.info(f"Image dimensions: {width}x{height}")

        # Check for symmetry if we have description
        symmetry_score = 7.0
        if description and 'symmetry' in description:
            symmetry_score = check_symmetry(image_path, description['symmetry'])

        # Simulate evaluation based on iteration
        if iteration_number == 1:
            # First iteration: moderate score with constructive feedback
            overall_score = 7.0
            shape_accuracy = 7.5
            proportion = 6.5
            resemblance = 7.0

            feedback = """First iteration analysis:
- Shape accuracy is good, curves are forming recognizable patterns
- Proportions could be improved - some components may need scaling adjustments
- Consider increasing detail or adjusting relative sizes of components
- Symmetry appears reasonable but could be refined"""

        elif iteration_number == 2:
            # Second iteration: improved score
            overall_score = 8.5
            shape_accuracy = 8.5
            proportion = 8.0
            resemblance = 9.0

            feedback = """Second iteration shows improvement:
- Shape accuracy has improved significantly
- Proportions are better balanced
- The drawing now more closely resembles the target
- Minor refinements could still be made to details"""

        else:
            # Third+ iteration: high score to indicate convergence
            overall_score = 9.0
            shape_accuracy = 9.0
            proportion = 9.0
            resemblance = 9.5

            # Adjust score based on symmetry if applicable
            if symmetry_score < 7.0:
                overall_score = max(8.0, overall_score - 0.5)

            feedback = """Final iteration analysis:
- Excellent shape accuracy achieved
- Proportions are well-balanced
- Strong resemblance to the intended object
- Drawing quality meets expectations"""

        # Log detailed scores
        logger.info(f"Evaluation scores - Overall: {overall_score}, "
                   f"Shape: {shape_accuracy}, Proportion: {proportion}, "
                   f"Resemblance: {resemblance}, Symmetry: {symmetry_score}")

        return overall_score, feedback

    except Exception as e:
        logger.error(f"Error evaluating image: {e}")
        # Return moderate score on error
        return 7.0, f"Evaluation encountered an error: {str(e)}"


def evaluate_with_vision_model(image_path: str, prompt_text: str) -> Tuple[float, str]:
    """
    Placeholder for future vision model integration.

    This function would call an actual vision AI model like:
    - Anthropic Claude with vision capabilities
    - OpenAI GPT-4V
    - A custom trained model

    Args:
        image_path: Path to the rendered image
        prompt_text: Original user prompt

    Returns:
        Tuple of (overall_score, feedback_text)
    """
    # TODO: Implement actual vision model integration
    # Example with Claude (when vision API is available):
    #
    # client = anthropic.Anthropic(api_key=ANTHROPIC_API_KEY)
    # with open(image_path, 'rb') as f:
    #     image_data = base64.b64encode(f.read()).decode()
    #
    # response = client.messages.create(
    #     model="claude-3-opus-20240229",  # or vision-enabled model
    #     max_tokens=500,
    #     messages=[{
    #         "role": "user",
    #         "content": [
    #             {
    #                 "type": "image",
    #                 "source": {
    #                     "type": "base64",
    #                     "media_type": "image/png",
    #                     "data": image_data
    #                 }
    #             },
    #             {
    #                 "type": "text",
    #                 "text": f"Evaluate this parametric curve drawing. "
    #                        f"The intended object was: '{prompt_text}'. "
    #                        f"Rate it 0-10 and provide specific feedback."
    #             }
    #         ]
    #     }]
    # )
    #
    # Parse response and extract score and feedback

    logger.warning("Vision model integration not yet implemented")
    return evaluate_image(image_path, prompt_text)


if __name__ == "__main__":
    # Test the evaluator
    print("Evaluator agent ready")
    print("Note: Currently using stub implementation")
    print("TODO: Integrate with actual vision AI model")
